[REFACTOR]: Command Type 경로 변경
