+++ apps/transcoder/src/main/java/com/ott/transcoder/command/Command.java
@@ -1,5 +1,7 @@
 package com.ott.transcoder.command;
 
+import com.ott.domain.ingest_command.domain.CommandType;
+
 public interface Command {
 
     CommandType getType();
