+++ apps/transcoder/src/main/java/com/ott/transcoder/job/JobOrchestrator.java
@@ -3,7 +3,7 @@ package com.ott.transcoder.job;
 import com.ott.domain.video_profile.domain.Resolution;
 import com.ott.transcoder.command.Command;
 import com.ott.transcoder.command.CommandExtractor;
-import com.ott.transcoder.command.CommandType;
+import com.ott.domain.ingest_command.domain.CommandType;
 import com.ott.transcoder.command.TranscodeCommand;
 import com.ott.transcoder.exception.TranscodeErrorCode;
 import com.ott.transcoder.exception.retryable.StorageException;
